+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Released on Wednesday, September 9 2026.
 - Added support for rounded corners
 - Added support for `text-shadow` declarations
 - Added support for `box-shadow` declarations
+- Added support for numeric sorted lists (e.g., using `ol`)
+- Added support for bullet point lists (e.g., using `ul`)
 
 # 0.3.0